Changes force_download.php to avoid proxying files
We've been proxying files through force_download.php so we can set headers to force downloading rather than opening in browser. But we only need to do that for a small subset of files. For S3 files, we can use presigned URLs to force download, and a simply query string parameter forces download of Vimeo CDN files. So we can save lots of bandwidth and time by only proxying files we _can't_ force download of.